Add security

+++ controllers/sauce.js
@@ -1,10 +1,15 @@
 const Sauce = require('../models/sauce');
+const fs = require('fs');
 
 // Get all sauces
 exports.readAllSauces = (req, res, next) => {
   Sauce.find()
     .then(allSauces => {
-      allSauces.map(sauce => sauce.imageUrl = `${req.protocol}://${req.get('host')}${sauce.imageUrl}`);
+      allSauces = allSauces.map(sauce => {
+        sauce.imageUrl = `${req.protocol}://${req.get('host')}${sauce.imageUrl}`
+        const links = linksHateoas(sauce._id)
+        return {...sauce._doc, links}
+      })
       res.status(200).json(allSauces);
     })
     .catch(error => res.status(500).json({error}));
@@ -15,7 +20,7 @@ exports.readOneSauce = (req, res, next) => {
   Sauce.findOne({_id: req.params.id})
     .then(sauce => {
       sauce.imageUrl = `${req.protocol}://${req.get('host')}${sauce.imageUrl}`;
-      res.status(200).json(sauce);
+      res.status(200).json(sauce, linksHateoas(sauce._id));
     })
     .catch(error => res.status(500).json({error}));
 };
@@ -23,15 +28,13 @@ exports.readOneSauce = (req, res, next) => {
 // Create a sauce
 exports.createSauce = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
-  delete sauceObject._id;
-  delete sauceObject._userId;
   const sauce = new Sauce({
     ...sauceObject,
     userId: req.auth.userId,
     imageUrl: `/images/${req.file.filename}`
   });
   sauce.save()
-    .then(() => res.status(201).json({message: 'Sauce created !'}))
+    .then((sauce) => res.status(201).json(sauce, linksHateoas(sauce._id)))
     .catch(error => res.status(401).json({error}));
 };
 
@@ -39,17 +42,22 @@ exports.createSauce = (req, res, next) => {
 exports.updateSauce = (req, res, next) => {
   const sauceObject = req.file ? {
     ...JSON.parse(req.body.sauce),
-    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    imageUrl: `/images/${req.file.filename}`
   } : { ...req.body };
 
-  delete sauceObject._userId;
   Sauce.findOne({_id: req.params.id})
     .then((sauce) => {
         if (sauce.userId != req.auth.userId) {
-            res.status(401).json({ message : 'Not authorized'});
+          res.status(401).json({ message : 'Not authorized'});
         } else {
-            Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
-            .then(() => res.status(200).json({message : 'Sauce modified!'}))
+          // Remove the old image from directory before adding the new one
+          const filename = sauce.imageUrl.split('/images/')[1];
+          if (sauceObject.imageUrl) { // Verify that there is an existing image before
+            fs.unlinkSync(`images/${filename}`);
+          }
+
+          Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
+            .then((sauce) => res.status(200).json(sauce, linksHateoas(sauce._id)))
             .catch(error => res.status(401).json({ error }));
         }
     })
@@ -58,9 +66,21 @@ exports.updateSauce = (req, res, next) => {
     });
 };
 
+// Delete a sauce
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({_id: req.params.id})
-    .then((deletedSauce) => res.status(200).json({message: 'Sauce deleted'}))
+  Sauce.findOne({_id: req.params.id})
+    .then(sauce => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message : 'Not authorized'});
+      } else {
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({_id: req.params.id})
+          .then((sauce) => res.status(200).json(sauce, linksHateoas(sauce._id)))
+          .catch(error => res.status(400).json({error}));
+        });
+      }
+    })
     .catch(error => res.status(400).json({error}));
 };
 
@@ -158,3 +178,43 @@ exports.likeSauce = (req, res, next) => {
     })
     .catch(error => res.status(400).json({error}));
 };
+
+/**
+ * Model of Hateoas links returned for all methods
+ */
+function linksHateoas(sauceId) {
+  
+  return [
+    {
+      rel: "createSauce",
+      method: "POST",
+      href: "http://localhost:3000/api/sauces",
+    },
+    {
+      rel: "updateSauce",
+      method: "PUT",
+      href: `http://localhost:3000/api/sauces/${sauceId}`,
+    },
+    {
+      rel: "readOneSauce",
+      method: "GET",
+      href: `http://localhost:3000/api/sauces/${sauceId}`,
+    },
+    {
+      rel: "readAllSauce",
+      method: "GET",
+      href: "http://localhost:3000/api/sauces",
+    },
+
+    {
+      rel: "deleteSauce",
+      method: "DELETE",
+      href: `http://localhost:3000/api/sauces/${sauceId}`,
+    },
+    {
+      rel: "likeSauce",
+      method: "POST",
+      href: `http://localhost:3000/api/sauces/${sauceId}/like`,
+    }
+  ];
+}
